==> 0_arabidopsis_interpolation.ipynb <==
import pandas as pd
import numpy as np
from scipy.interpolate import interp1d # for interpolation

# create function to return distances of polyline vectors

def get_distance(x, y):

    d = np.cumsum(np.sqrt( np.ediff1d(x)**2 + np.ediff1d(y)**2 ))
    d = d[-1]

    return d

# create empty list to store overall distance of polyline for each leaf, 1 through 11

leaf_distances = []

# iterate over each file for each leaf and store overall distance

for leaf in range(11):
    
    print(leaf)
    
    leaf_number = str(leaf + 1) # add one because leaf index starts on 1 not 0
    
    leaf_number = leaf_number.zfill(2) # file name numbers are 2 places with a leading zero
    
    file_name = "./leaf" + leaf_number + ".txt" # create file name

    leaf_data = np.loadtxt(file_name) # load in data for the current leaf
    
    # calculate distances for leaf outline polyline and append to list

    leaf_distances.append(get_distance(leaf_data[:,0], leaf_data[:,1]))
    
# check leaf distance successfully calculated

print(leaf_distances)

# function to retrieve iterpolated points

def interpolation(x, y, number):

    distance = np.cumsum(np.sqrt( np.ediff1d(x, to_begin=0)**2 + np.ediff1d(y, to_begin=0)**2 ))
    distance = distance/distance[-1]

    fx, fy = interp1d( distance, x ), interp1d( distance, y )

    alpha = np.linspace(0, 1, number)
    x_regular, y_regular = fx(alpha), fy(alpha)
    
    return x_regular, y_regular

n_pts = 100 # select number of points to interpolate for each leaf

interpolated_points = [] # list of lists with points for each leaf

for leaf in range(11):
    
    print(leaf)
    
    leaf_number = str(leaf + 1) # add one because leaf index starts on 1 not 0
    
    leaf_number = leaf_number.zfill(2) # file name numbers are 2 places with a leading zero
    
    file_name = "./leaf" + leaf_number + ".txt" # create file name

    leaf_data = np.loadtxt(file_name) # load in data for the current leaf
    
    # perform interpolation
    
    xpts, ypts = interpolation(leaf_data[:,0], leaf_data[:,1], n_pts)
    
    leaf_vals = list(zip(xpts, ypts)) # points in tuples of form (x1, y1), etc.
    leaf_vals = list(sum(leaf_vals, ())) # points converted to an x1, y1, x2, y2 ... xn, yn list
    
    interpolated_points.append(leaf_vals) # append points to overall list

print(len(interpolated_points)) # check list length

df = pd.DataFrame(interpolated_points) # convert to a pd.df

df.head() # check df
df.tail() # check df

df.to_csv('./interpolated_points.txt', sep='\t') # write out file
