==> 2_arabidopsis_modeling_landmarks.ipynb <==
import numpy as np
import matplotlib.pyplot as plt
from scipy.spatial import distance
%matplotlib inline


xdata = np.loadtxt("./xvals_for_modeling.csv", delimiter=",") # load in x vals, rows = leaf, cols = landmarks 1-100
ydata = np.loadtxt("./yvals_for_modeling.csv", delimiter=",") # load in y vals, rows = leaf, cols = landmarks 1-100


# create a loop to fit a polynomial for each of 100 landmarks xvals

order = 3 # polynomial degree

model_no = 100 # number of positions to model leaf development

dev_xvals = [] # store a list of arrays, each array model for each of 100 landmarks

for i in range(100): # for each of 100 landmarks
    
    xvals = list(range(11)) # model is across 11 leaves
    yvals = xdata[:,i] # isolate the landmark for the current loop iteration
    params = np.polyfit(xvals, yvals, order) # find poly params
    func = np.poly1d(params) # create poly function
    current_landmark_xvals = func(np.linspace(0,10,model_no)) # calculate model values
    dev_xvals.append(current_landmark_xvals) # append model values
    
dev_xvals = np.array(dev_xvals) # save as an array, landmark by developmental stage

# create a loop to fit a polynomial for each of 100 landmarks yvals

dev_yvals = [] # store a list of arrays, each array model for each of 100 landmarks

for i in range(100): # for each of 100 landmarks
    
    xvals = list(range(11)) # model is across 11 leaves
    yvals = ydata[:,i] # isolate the landmark for the current loop iteration
    params = np.polyfit(xvals, yvals, order) # find poly params
    func = np.poly1d(params) # create poly function
    current_landmark_yvals = func(np.linspace(0,10,model_no)) # calculate model values
    dev_yvals.append(current_landmark_yvals) # append model values
    
dev_yvals = np.array(dev_yvals) # save as an array, landmark by developmental stage

# loop to plot out leaves as a growing rosette using the golden angle

golden_ang = 137.5077640500378546463487*(3.14159/180)
limit=500


for j in range(1, 1000, 4):
    
    iteration = j

    new_leaf_num = iteration//10 + 1
    old_leaf_num = new_leaf_num - 9

    leaves = np.arange(old_leaf_num, new_leaf_num + 1)

    thetas = []

    for i in leaves:

        curr_theta = golden_ang*i
        thetas.append(curr_theta)

    for i in range(10):

        radian_ang = thetas[i]
        stage = ((9-i)*10) + (iteration%10)


        rotated_xvals = dev_xvals[:,stage]*np.cos(radian_ang) - dev_yvals[:,stage]*np.sin(radian_ang)
        rotated_yvals = dev_yvals[:,stage]*np.cos(radian_ang) + dev_xvals[:,stage]*np.sin(radian_ang)

        plt.style.use('dark_background')
        plt.plot(rotated_xvals, rotated_yvals, c="magenta")
        plt.gca().set_aspect("equal")
        plt.axis("off")
        plt.xlim(-limit,limit)
        plt.ylim(-limit,limit)
        
    plt.savefig("./temp/x_" + str(j) + ".jpg")
    plt.close()
    plt.clf()
